@@ -25,73 +25,66 @@
   "license": "ISC",
   "homepage": "",
   "dependencies": {
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.10.5",
-    "@babel/node": "^7.10.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-decorators": "^7.10.5",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
+    "@babel/cli": "^7.12.1",
+    "@babel/core": "^7.12.3",
+    "@babel/node": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
     "@reduxjs/toolkit": "^1.4.0",
-    "axios": "^0.19.2",
-    "babel-loader": "^8.1.0",
-    "bootstrap": "^4.5.0",
+    "axios": "^0.20.0",
+    "bootstrap": "^4.5.3",
     "classnames": "^2.2.6",
     "core-js": "^3.6.5",
-    "css-loader": "^3.6.0",
     "cssnano": "^4.1.10",
-    "debug": "^4.1.1",
+    "debug": "^4.2.0",
     "dotenv": "^8.2.0",
-    "faker": "^4.1.0",
-    "fastify": "^3.0.3",
-    "fastify-formbody": "^4.0.3",
-    "fastify-secure-session": "^1.2.0",
+    "faker": "^5.1.0",
+    "fastify": "^3.7.0",
+    "fastify-formbody": "^5.0.0",
+    "fastify-secure-session": "^2.3.0",
     "fastify-sensible": "^3.0.1",
-    "fastify-static": "^3.1.0",
-    "formik": "^2.1.5",
+    "fastify-static": "^3.3.0",
+    "formik": "^2.2.0",
     "jquery": "^3.5.1",
     "js-cookie": "^2.2.1",
-    "lodash": "^4.17.19",
-    "mini-css-extract-plugin": "^0.9.0",
+    "lodash": "^4.17.20",
     "node-sass": "^4.14.1",
-    "point-of-view": "^4.2.0",
+    "point-of-view": "^4.6.0",
     "popper.js": "^1.16.1",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.7.0",
     "pug": "^3.0.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.2.2",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.0",
+    "react-bootstrap": "^1.4.0",
+    "react-dom": "^17.0.0",
     "react-redux": "^7.2.1",
-    "redux": "^4.0.5",
-    "redux-actions": "^2.6.5",
-    "redux-bootstrap": "^1.3.0",
-    "redux-thunk": "^2.3.0",
-    "regenerator-runtime": "^0.13.5",
-    "sass-loader": "^9.0.2",
+    "regenerator-runtime": "^0.13.7",
     "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.1",
-    "style-loader": "^1.2.1",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "socket.io-client": "^3.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.4",
+    "@types/jest": "^26.0.15",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.1.0",
-    "eslint": "^7.4.0",
+    "babel-jest": "^26.6.0",
+    "babel-loader": "^8.1.0",
+    "css-loader": "^5.0.0",
+    "eslint": "^7.11.0",
     "eslint-config-airbnb": "^18.2.0",
-    "eslint-import-resolver-webpack": "^0.12.2",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^23.18.0",
+    "eslint-import-resolver-webpack": "^0.13.0",
+    "eslint-plugin-functional": "^3.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jest": "^24.1.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react-hooks": "^4.0.8",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-functional": "^3.0.1",
-    "jest": "^26.1.0",
-    "jest-cli": "^26.1.0",
-    "nodemon": "^2.0.4"
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "jest": "^26.6.0",
+    "jest-cli": "^26.6.0",
+    "mini-css-extract-plugin": "^1.1.1",
+    "nodemon": "^2.0.6",
+    "postcss-loader": "^4.0.4",
+    "postcss-preset-env": "^6.7.0",
+    "sass-loader": "^10.0.3",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.1.3",
+    "webpack-cli": "^4.1.0",
+    "webpack-dev-server": "^3.11.0"
   }
 }
