@@ -1,7 +1,5 @@
 module.exports = {
-  plugins: {
-    'postcss-import': {},
-    'postcss-preset-env': {},
-    cssnano: {},
-  },
+  plugins: [
+    'postcss-preset-env',
+  ],
 };
