@@ -1,8 +1,6 @@
 // @ts-check
 
-import 'core-js/stable';
 import 'regenerator-runtime/runtime';
-
 import path from 'path';
 import Pug from 'pug';
 import socket from 'socket.io';
@@ -37,7 +35,7 @@ const setUpStaticAssets = (app) => {
 };
 
 export default (options) => {
-  const app = fastify();
+  const app = fastify({ logger: true, prettyPrint: true });
 
   setUpViews(app);
   setUpStaticAssets(app);
